+++ app/assets/scripts/cards/ad_unit.js
@@ -14,12 +14,18 @@ function( angular ) {
                 _data = config._data || (config._data = {
                     hasPlayed: false,
                     companion: null,
+                    tracking: {
+                        clickFired: false,
+                        countFired: false
+                    },
                     modules: {
                         displayAd: {
                             active: false
                         }
                     }
-                });
+                }),
+                lastTime = null,
+                elapsedTime = 0;
 
             function playerReady(player) {
                 self.player = player;
@@ -100,6 +106,40 @@ function( angular ) {
                         deactivateCard();
                     }
                 });
+
+                // If it's a sponsored card, set up handlers to fire AdCount and Click pixels
+                if (config.campaign) {
+                    // Fire the Click pixel after the first play
+                    if (config.campaign.clickUrl && !_data.tracking.clickFired) {
+                        player.once('play', function() {
+                            _data.tracking.clickFired = true;
+                            c6ImagePreloader.load([config.campaign.clickUrl]);
+                        });
+                    }
+
+                    // Fire the AdCount pixel after minViewTime, by tracking the elapsed time
+                    if (config.campaign.countUrl && config.campaign.minViewTime &&
+                                                    !_data.tracking.countFired) {
+                        player.on('timeupdate', function fireMinViewPixel() {
+                            if (lastTime === null) {
+                                lastTime = player.currentTime;
+                                return;
+                            }
+
+                            // if diff > 1 sec, it's probably a skip, and don't increment elapsed
+                            if (Math.abs(player.currentTime - lastTime) <= 1) {
+                                elapsedTime += player.currentTime - lastTime;
+                            }
+                            lastTime = player.currentTime;
+
+                            if (elapsedTime >= config.campaign.minViewTime && !_data.tracking.countFired) {
+                                _data.tracking.countFired = true;
+                                c6ImagePreloader.load([config.campaign.countUrl]);
+                                player.removeListener('timeupdate', fireMinViewPixel);
+                            }
+                        });
+                    }
+                }
             }
 
             function playerInit($event, player) {